test(grid): move Kbnav test in a separeta file #2975
